moved away from spatialbuilder to scopes to allow projects to define their own builders
