+++ phpstan.neon
@@ -1,5 +1,5 @@
 includes:
-  - ./vendor/nunomaduro/larastan/extension.neon
+  - ./vendor/larastan/larastan/extension.neon
 parameters:
   paths:
     - src
