+++ tests/TestModels/TestPlace.php
@@ -11,7 +11,6 @@ use MatanYadaev\EloquentSpatial\Objects\MultiPoint;
 use MatanYadaev\EloquentSpatial\Objects\MultiPolygon;
 use MatanYadaev\EloquentSpatial\Objects\Point;
 use MatanYadaev\EloquentSpatial\Objects\Polygon;
-use MatanYadaev\EloquentSpatial\SpatialBuilder;
 use MatanYadaev\EloquentSpatial\Tests\TestFactories\TestPlaceFactory;
 use MatanYadaev\EloquentSpatial\Traits\HasSpatial;
 
@@ -28,8 +27,6 @@ use MatanYadaev\EloquentSpatial\Traits\HasSpatial;
  * @property Point|null $centroid
  * @property Point|null $centroid_alias
  * @mixin Model
- *
- * @method static SpatialBuilder<TestPlace> query()
  */
 class TestPlace extends Model
 {
